Smart Contract Completed

+++ src/trade-offer/trade_offer.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, UnauthorizedException } from "@nestjs/common";
+import { ForbiddenException, Injectable, UnauthorizedException } from "@nestjs/common";
 import { UserRepository } from "src/users/user.repository";
 import { TradeOfferRepository } from "./trade_offer.repository";
 import { ProductRepository } from "src/products/product.repository";
 import { ProductStatus } from "src/types/product-status";
 import { TradeOfferStatus } from "src/types/trade-offer-status";
+import { Product } from "src/products/entities/product.entity";
 
 @Injectable()
 export class TradeOfferService {
@@ -55,11 +56,18 @@ export class TradeOfferService {
         const user = await this.userRepository.findOne(username);
         const to = await this.tradeOfferRepository.findOne(id);
 
+        const product = to.product;
         if (to.buyer.id !== user.id) {
             throw new UnauthorizedException("You have not permission to remove this TO");
         }
 
-        return await this.tradeOfferRepository.remove(to);
+        if (product.status === ProductStatus.FINDING || product.status === ProductStatus.MATCHED) {
+            if (product.status === ProductStatus.MATCHED)
+                product.status = ProductStatus.FINDING;
 
+            return await this.tradeOfferRepository.remove(to);
+        } else {
+            throw new ForbiddenException("Can't remove Trade Offer because of product status.");
+        }
     }
 }
